% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat.R
\name{as.character.hd_Dat}
\alias{as.character.hd_Dat}
\title{#' @method vec_cast.character hd_Dat
#' @export
vec_cast.character.hd_Dat <- function(x, to, ...) Dat_show(x)}
\usage{
\method{as.character}{hd_Dat}(x)
}
\description{
#' @method vec_cast.character hd_Dat
#' @export
vec_cast.character.hd_Dat <- function(x, to, ...) Dat_show(x)
}
