% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat.R
\name{as_Dat}
\alias{as_Dat}
\title{#' @method vec_cast.date hd_Dat
#' @export
vec_cast.date.hd_Dat <- function(x, to, ...) as.character(vctrs::new_date(vctrs::vec_data(x)))}
\usage{
as_Dat(x)
}
\description{
#' @method vec_cast.date hd_Dat
#' @export
vec_cast.date.hd_Dat <- function(x, to, ...) as.character(vctrs::new_date(vctrs::vec_data(x)))
}
