% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frType.R
\name{as_frType}
\alias{as_frType}
\title{frType Vectors}
\usage{
as_frType(x)
}
\arguments{
\item{x}{object to be coerced}
}
\value{
attempts to coerce its argument to frType type
}
\description{
coerces its argument to a frType. It is an abbreviated form of frType.
}
\examples{

some_chr_value <- "Cat"
class(some_chr_value)

some_frt_value <- as_frType(some_chr_value)
class(some_frt_value)

}
