% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTypes.R
\name{as_hdType}
\alias{as_hdType}
\title{hdType Vectors}
\usage{
as_hdType(x)
}
\arguments{
\item{x}{object to be coerced}
}
\value{
attempts to coerce its argument to hdType type
}
\description{
coerces its argument to a hdType. It is an abbreviated form of hdType.
}
\examples{

some_chr_value <- "Cat"
class(some_chr_value)

some_hdt_value <- as_hdType(some_chr_value)
class(some_hdt_value)

}
