% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castCtypes.R
\name{cast_ctypes}
\alias{cast_ctypes}
\title{cast_ctypes
Convert Ctypes}
\usage{
cast_ctypes(data, from, to)
}
\arguments{
\item{data}{frame or a fringe.}

\item{from}{string of Ctypes.}

\item{to}{string of Ctypes.}
}
\value{
data frame or a fringe according to input.
}
\description{
cast_ctypes
Convert Ctypes
}
\examples{
cast_ctypes(iris,"","")
}
