% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castCtypes.R
\name{cast_Da2Ca_weeks}
\alias{cast_Da2Ca_weeks}
\title{cast_Da2Ca_weeks : title.
Convert Data type to Category weeks}
\usage{
cast_Da2Ca_weeks(data, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
cast_Da2Ca_weeks : title.
Convert Data type to Category weeks
}
\section{ftypes}{
 Da-Nu
}
\examples{
add(1, 1)
add(10, 1)
}

