% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_hdtypes.R
\name{cast_hdtypes}
\alias{cast_hdtypes}
\title{cast_hdtypes
Convert hdTypes}
\usage{
cast_hdtypes(data, from, to)
}
\arguments{
\item{data}{frame or a fringe.}

\item{from}{string of hdtypes.}

\item{to}{string of hdtypes.}
}
\value{
data frame or a fringe according to input.
}
\description{
cast_hdtypes
Convert hdTypes
}
\examples{
cast_hdtypes(iris,"","")
}
