% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frType.R
\name{expand_frGroup}
\alias{expand_frGroup}
\title{frType Vectors}
\usage{
expand_frGroup(frGroup)
}
\arguments{
\item{frGroup}{a grouped frType object}
}
\value{
a string value showing all the frType values of an object
}
\description{
Reverses the effect of \code{\link[=get_frGroup]{get_frGroup()}} and split every single frType from an object.
}
\examples{

x <- c("Cat-Num-Cat")
fr <- frType(x)
grouped_fr <- get_frGroup(fr)
expand_frGroup(grouped_fr)

}
