% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fringe.R
\name{fringe}
\alias{fringe}
\title{Create a Fringe data frame}
\usage{
fringe(
  x = new_data_frame(),
  frtype = NULL,
  dic = NULL,
  name = NULL,
  description = NULL,
  slug = NULL,
  meta = NULL,
  ...
)
}
\arguments{
\item{x}{A data frame}

\item{frtype}{The type of fringe to create}

\item{dic}{a custom variable dictionary can be added. \code{\link[=create_dic]{create_dic()}} can help you with that.}

\item{name}{a custom name can be added}

\item{slug}{a custom slug can be added. If not, fringe will try creating one.}

\item{meta}{Custom Metadata can be added}

\item{nam}{a custom description can be added}
}
\value{
A Fringe object
}
\description{
Create a Fringe object from a data frame. The main value of a fringe is its metadata. When creating it, fringe will add to the data frame the following information:
\itemize{
\item data: original data frame data. When it is created, the fringe will convert original variable R types onto homodatum ones (Num, Cat, Pct, etc. -see \code{\link[=available_hdTypes]{available_hdTypes()}})
\item dic: A diccionary is created with three variable characteristics: id, label and hdType
\item frType: Shows all variable types based in homodatum schema
\item group: A grouped view of frType
\item name: Name for the fringe data frame, setted on \emph{name} argument
\item description: Description for the fringe data frame, setted on \emph{description} argument
\item slug: a custom slug can be added
\item stats: Depending on the variable type given by homodatum, the fringe will generate different kind of statistics: nrow, ncol, n_unique, n_na, pct_na, min, max
}
}
\examples{
fringe(mtcars, frtype = "Num", name = "MTCars")

}
