% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frType.R
\name{frType}
\alias{frType}
\title{frType Vectors}
\usage{
frType(x = character())
}
\arguments{
\item{x}{object to be created or coerced}
}
\value{
returns a frType value
}
\description{
Creates or coerces objects of type "frType"
}
\examples{

frType("Cat")

}
