% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frType.R
\name{frType_group}
\alias{frType_group}
\title{frType Vectors}
\usage{
frType_group(x)
}
\arguments{
\item{x}{several (more than one) available frTypes values}
}
\value{
a grouped view of given frTypes values
}
\description{
a grouped way of reading frTypes values
}
\examples{

x <- c("Cat-Num-Cat")
fr <- frType(x)
frType_group(fr)

}
