% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frType.R
\name{frType_hdTypes}
\alias{frType_hdTypes}
\title{frType Vectors}
\usage{
frType_hdTypes(x, chr = FALSE)
}
\arguments{
\item{x}{an available frType value}
}
\value{
an hdType value
}
\description{
convert frTypes value(s) into hdType
}
\examples{

x <- frType("Cat")
class(x)

x_hdt <- frType_hdTypes(x)
class(x_hdt)

}
