% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frType.R
\name{frType_str}
\alias{frType_str}
\title{frType Vectors}
\usage{
frType_str(x)
}
\arguments{
\item{x}{an available frType value, frType dataframe or hd_tbl dataframe}
}
\value{
a character value
}
\description{
convert frTypes value(s) into character
}
\examples{

x <- frType("Cat")
class(x)

x_chr <- frType_str(x)
class(x_chr)

}
