% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frType.R
\name{get_frGroup}
\alias{get_frGroup}
\title{frType Vectors}
\usage{
get_frGroup(frType_str)
}
\arguments{
\item{frType_str}{a string value showing a grouped frTypes view}
}
\value{
a grouped view of given frTypes values
}
\description{
a grouped way of reading frTypes values
}
\examples{

x <- c("Cat-Num-Cat")
fr <- frType(x)
get_frGroup(fr)

}
