% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_hdType.R
\name{guess_hdType}
\alias{guess_hdType}
\title{hdType}
\usage{
guess_hdType(v)
}
\arguments{
\item{v}{a value for which you want to evaluate}
}
\value{
An available_hdType value
}
\description{
evaluates which of the available_hdTypes is the object
}
\examples{

# Num hdType
value <- c("1",NA,"2")
guess_hdType(value)

# Cat hdType
value <- c("MSPS-CD-166-2020", "003-2020", "0811 - 2020")
guess_hdType(value)

# Pct hdType
value <- c(0.3, 0.4, 1)
guess_hdType(value)


}
