% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTypes.R
\name{is_any_hdType}
\alias{is_any_hdType}
\title{hdType Vectors}
\usage{
is_any_hdType(x)
}
\arguments{
\item{x}{hdType object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is one of the \code{\link[=available_hdTypes]{available_hdTypes()}} or not.
}
\description{
test for objects of type "hdType"
}
\examples{

some_cat_value <- Cat("value1")
is_any_hdType(some_cat_value)

}
