% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cat.R
\name{is_Cat}
\alias{is_Cat}
\title{Category Vectors}
\usage{
is_Cat(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hd_Cat or not.
}
\description{
test for objects of type "hd_Cat"
}
\examples{
some_value <- Cat(c("Cat1", "Cat2"))
is_Cat(some_value)

}
