% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat.R
\name{is_Dat}
\alias{is_Dat}
\title{Date Vectors}
\usage{
is_Dat(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hd_Dat or not.
}
\description{
test for objects of type "hd_Dat"
}
\examples{
x <-   c("2000-12-04","2010-20-04")
dat <- Dat(x, format = "\%Y-\%d-\%m")
is_Dat(dat)

}
