% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fringe.R
\name{is_fringe}
\alias{is_fringe}
\title{Fringe data frame}
\usage{
is_fringe(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type fringe or not.
}
\description{
test for objects of type "fringe"
}
\examples{
some_df <- fringe(mtcars)
is_fringe(some_df)

}
