% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frType.R
\name{is_frType}
\alias{is_frType}
\title{frType Vectors}
\usage{
is_frType(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type frType or not.
}
\description{
Creates or test for objects of type "frType"
}
\examples{

value <- frType("Cat")
is_frType(value)

}
