% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTypes.R
\name{is_hdType}
\alias{is_hdType}
\title{hdType Vectors}
\usage{
is_hdType(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdType or not.
}
\description{
test for objects of type "hdType"
}
\examples{

some_value <- hdType("Cat")
is_hdType(some_value)

}
