% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num.R
\name{is_Num}
\alias{is_Num}
\title{Numeric Vectors}
\usage{
is_Num(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hd_Num or not.
}
\description{
test for objects of type "hd_Num"
}
\examples{
x <-   c(1, 54, 678, NA, NA)
num <- Num(x)
is_Num(num)

}
