% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pct.R
\name{is_Pct}
\alias{is_Pct}
\title{Percentage Vectors}
\usage{
is_Pct(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hd_Pct or not.
}
\description{
test for objects of type "hd_Pct"
}
\examples{
pct_value <- Pct(c("1","0.2"))
is_Pct(pct_value)

}
