% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yea.R
\name{is_Yea}
\alias{is_Yea}
\title{Year Vectors}
\usage{
is_Yea(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hd_Yea or not.
}
\description{
test for objects of type "hd_Yea"
}
\examples{
x <- c(1800:1803, NA)
year_values <- Yea(x)
is_Yea(year_values)

}
