% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num.R
\name{Num}
\alias{Num}
\title{Numeric Vectors}
\usage{
Num(x = double())
}
\arguments{
\item{x}{object to be created as Num type}
}
\description{
Creates objects of type "hd_Num". hd_Num objects contain a "stats" attribute, with the following information: unique values (n_unique), NA values (n_na), percentage weight of NA values (pct_na), lowest value (min) and highest value (max).
}
\examples{
x <-   c(1, 54, 678, NA, NA)
num <- Num(x)
num
class(num)
attr(num, "stats")

}
