% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pct.R
\name{Pct}
\alias{Pct}
\title{Percetage Vectors}
\usage{
Pct(x = double())
}
\arguments{
\item{x}{object to be created as Num type}
}
\description{
Creates objects of type "hd_Pct". hd_Pct format accepts anything coercible from double(). It also recognizes number with percentage symbol in a character value, as it could be "20\%".
}
\examples{

## Example 1
pct_value <- Pct(c("1","0.2"))
pct_value
class(pct_value)

## Example 2
pct_value <- Pct(c("30 (\%)", "(\%)200", NA))
pct_value
class(pct_value)

}
