% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bin.R, R/Cat.R, R/Dat.R, R/homodatum-package.R
\name{vec_ptype2.hd_Bin}
\alias{vec_ptype2.hd_Bin}
\alias{vec_cast.hd_Bin}
\alias{vec_ptype2.hd_Cat}
\alias{vec_cast.hd_Cat}
\alias{vec_ptype2.hd_Dat}
\alias{vec_cast.hd_Dat}
\alias{vctrs-compat}
\title{vctrs compatibility functions for math, casting, coercion, and comparison}
\usage{
\method{vec_ptype2}{hd_Bin}(x, y, ...)

\method{vec_cast}{hd_Bin}(x, to, ...)

\method{vec_ptype2}{hd_Cat}(x, y, ...)

\method{vec_cast}{hd_Cat}(x, to, ...)

\method{vec_ptype2}{hd_Dat}(x, y, ...)

\method{vec_cast}{hd_Dat}(x, to, ...)
}
\arguments{
\item{x, y}{Vectors.}

\item{...}{Arguments passed on to further methods.}

\item{to}{Type to cast to.}

\item{op}{Arithmetic operation.}

\item{.fn}{Mathematical function.}

\item{.x}{Vector.}
}
\description{
These functions are the extensions that allow debkeepr vectors to
work with vctrs.
}
