% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat.R
\name{vec_cast.date.hd_Dat}
\alias{vec_cast.date.hd_Dat}
\title{#' @method vec_cast.date hd_Dat
#' @export}
\usage{
\method{vec_cast}{date.hd_Dat}(x, to, ...)
}
\description{
#' @method vec_cast.date hd_Dat
#' @export
}
