% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTypes.R
\name{which_hdType}
\alias{which_hdType}
\title{hdType Vectors}
\usage{
which_hdType(x)
}
\arguments{
\item{x}{hdType object to be coerced or tested}
}
\value{
returns the name of the hdType value. You can see the valid hdTypes with \code{\link[=available_hdTypes]{available_hdTypes()}}
}
\description{
Detect which hdType is the value.
}
\examples{

some_cat_value <- Cat("value1")
which_hdType(some_cat_value)

}
