% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homodatum-package.R
\docType{package}
\name{homodatum-package}
\alias{homodatum}
\alias{homodatum-package}
\title{homodatum: Estimate Gender from Names in Spanish and Portuguese}
\description{
Data types for humans.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/homodatum}
  \item Report bugs at \url{https://github.com/datasketch/homodatum/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
